+++ .eslintignore
@@ -1,1 +1,1 @@
-/lib/*
+/src/lib/*
