+++ src/app/components/characters/characters.scss
