+++ src/app/components/details/details.pug
