+++ src/app/components/details/details.scss
