+++ src/index.js
@@ -1,4 +1,4 @@
-import AppComponent from './app';
+import AppComponent from './app/app';
 
 const app = new AppComponent();
 app.createComponent();
